add build amd64

+++ download/download-sdk.sh
@@ -6,20 +6,27 @@ set -eux
 
 pushd /
 
-sudo wget -cO sdk.tar.gz https://repo.huaweicloud.com/openharmony/os/$V_SDK/ohos-sdk-windows_linux-public.tar.gz
-sudo mkdir -p sdk
-sudo tar -C sdk -zxf sdk.tar.gz
-sudo rm sdk.tar.gz
+if [ ! -d sdk ]; then
+  echo "Downloading sdk..."
+  sudo wget -cO sdk.tar.gz https://repo.huaweicloud.com/openharmony/os/$V_SDK/ohos-sdk-windows_linux-public.tar.gz
+  sudo mkdir -p sdk
+  sudo tar -C sdk -zxf sdk.tar.gz
+  sudo rm sdk.tar.gz
 
-cd sdk
-sudo rm -rf windows/
+  cd sdk
+  sudo rm -rf windows/
 
-# Extract NDK
-cd linux
-for i in *.zip
-do
-  sudo unzip -q $i
-  sudo rm $i
-done
+  # Extract NDK
+  cd linux
+  for i in *.zip
+  do
+    sudo unzip -q $i
+    sudo rm $i
+  done
+else
+  echo "dav1d already exists, skipping."
+fi
+
+
 
 popd
